chore: renamed logging to avoid conflict with py logging packagpkg
